Block code mk1
First iteration of movement and interaction with the different types of blocks. Implemented functions for creating new blocks, moving blocks left, right and down (current iteration not entirely optimised with side detection). Other functions include rotate which is able to rotate the block clockwise as well as other functions with the purpose of limiting repeating code.